Run ivy-ngcc on npm postinstall
It seems that Angular CLI runs the ngcc only on the root entrypoints
of the npm modules when performing a build. This leads to compile
errors because NgModules from deep imports can not be found,
e.g. HttpClientModule from '@angular/common/http' or all CDK modules.

As a workaround run ngcc on npm postinstall to compile/transform all
modules.